Done: 예약 - 스케줄 PC 화면 구현을 위해 컴포넌트 분리
